+++ lesson-7/src/firebase/config.js
