+++ lesson-7/vite.config.js
